requirements.txt
Flask
numpy<2
opencv-python
torch
torchvision
ultralytics
Pillow
tqdm
pillow_heif
timm
matplotlib